+++ library/LocalRegions/TriExp.h
@@ -61,14 +61,14 @@ public:
 
     LOCAL_REGIONS_EXPORT TriExp(const TriExp &T);
 
-    LOCAL_REGIONS_EXPORT ~TriExp();
+    LOCAL_REGIONS_EXPORT virtual ~TriExp() override = default;
 
 protected:
     //-------------------------------
     // Integration Methods
     //-------------------------------
     LOCAL_REGIONS_EXPORT virtual NekDouble v_Integral(
-        const Array<OneD, const NekDouble> &inarray);
+        const Array<OneD, const NekDouble> &inarray) override;
 
     //----------------------------
     // Differentiation Methods
@@ -76,173 +76,168 @@ protected:
     LOCAL_REGIONS_EXPORT virtual void v_PhysDeriv(
         const Array<OneD, const NekDouble> &inarray,
         Array<OneD, NekDouble> &out_d0, Array<OneD, NekDouble> &out_d1,
-        Array<OneD, NekDouble> &out_d2 = NullNekDouble1DArray);
+        Array<OneD, NekDouble> &out_d2 = NullNekDouble1DArray) override;
     LOCAL_REGIONS_EXPORT virtual void v_PhysDeriv(
         const int dir, const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, NekDouble> &outarray) override;
     LOCAL_REGIONS_EXPORT virtual void v_PhysDirectionalDeriv(
         const Array<OneD, const NekDouble> &inarray,
         const Array<OneD, const NekDouble> &direction,
-        Array<OneD, NekDouble> &out);
+        Array<OneD, NekDouble> &out) override;
 
     //---------------------------------------
     // Transforms
     //---------------------------------------
     LOCAL_REGIONS_EXPORT virtual void v_FwdTrans(
         const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, NekDouble> &outarray) override;
     LOCAL_REGIONS_EXPORT virtual void v_FwdTransBndConstrained(
         const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, NekDouble> &outarray) override;
 
     //---------------------------------------
     // Inner product functions
     //---------------------------------------
     LOCAL_REGIONS_EXPORT virtual void v_IProductWRTBase(
         const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, NekDouble> &outarray) override;
     LOCAL_REGIONS_EXPORT virtual void v_IProductWRTDerivBase(
         const int dir, const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, NekDouble> &outarray) override;
     LOCAL_REGIONS_EXPORT virtual void v_IProductWRTBase_SumFac(
         const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray, bool multiplybyweights = true);
-    LOCAL_REGIONS_EXPORT virtual void v_IProductWRTBase_MatOp(
-        const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, NekDouble> &outarray,
+        bool multiplybyweights = true) override;
     LOCAL_REGIONS_EXPORT virtual void v_IProductWRTDerivBase_SumFac(
         const int dir, const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, NekDouble> &outarray) override;
     LOCAL_REGIONS_EXPORT virtual void v_AlignVectorToCollapsedDir(
         const int dir, const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, Array<OneD, NekDouble>> &outarray);
-
-    LOCAL_REGIONS_EXPORT virtual void v_IProductWRTDerivBase_MatOp(
-        const int dir, const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, Array<OneD, NekDouble>> &outarray) override;
 
     LOCAL_REGIONS_EXPORT virtual void v_IProductWRTDirectionalDerivBase(
         const Array<OneD, const NekDouble> &direction,
         const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, NekDouble> &outarray) override;
 
     LOCAL_REGIONS_EXPORT virtual void v_IProductWRTDirectionalDerivBase_SumFac(
         const Array<OneD, const NekDouble> &direction,
         const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, NekDouble> &outarray) override;
 
     LOCAL_REGIONS_EXPORT virtual void v_NormVectorIProductWRTBase(
         const Array<OneD, const NekDouble> &Fx,
         const Array<OneD, const NekDouble> &Fy,
         const Array<OneD, const NekDouble> &Fz,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, NekDouble> &outarray) override;
 
     LOCAL_REGIONS_EXPORT virtual void v_NormVectorIProductWRTBase(
         const Array<OneD, const Array<OneD, NekDouble>> &Fvec,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, NekDouble> &outarray) override;
 
     //---------------------------------------
     // Evaluation functions
     //---------------------------------------
     LOCAL_REGIONS_EXPORT virtual StdRegions::StdExpansionSharedPtr v_GetStdExp(
-        void) const;
+        void) const override;
 
     LOCAL_REGIONS_EXPORT virtual StdRegions::StdExpansionSharedPtr v_GetLinStdExp(
-        void) const;
+        void) const override;
 
     LOCAL_REGIONS_EXPORT virtual void v_GetCoord(
         const Array<OneD, const NekDouble> &Lcoords,
-        Array<OneD, NekDouble> &coords);
+        Array<OneD, NekDouble> &coords) override;
     LOCAL_REGIONS_EXPORT virtual void v_GetCoords(
         Array<OneD, NekDouble> &coords_1, Array<OneD, NekDouble> &coords_2,
-        Array<OneD, NekDouble> &coords_3);
+        Array<OneD, NekDouble> &coords_3) override;
     LOCAL_REGIONS_EXPORT virtual NekDouble v_StdPhysEvaluate(
         const Array<OneD, const NekDouble> &Lcoord,
-        const Array<OneD, const NekDouble> &physvals);
+        const Array<OneD, const NekDouble> &physvals) override;
     LOCAL_REGIONS_EXPORT virtual NekDouble v_PhysEvaluate(
         const Array<OneD, const NekDouble> &coord,
-        const Array<OneD, const NekDouble> &physvals);
+        const Array<OneD, const NekDouble> &physvals) override;
     LOCAL_REGIONS_EXPORT virtual void v_GetTracePhysVals(
         const int edge, const StdRegions::StdExpansionSharedPtr &EdgeExp,
         const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray, StdRegions::Orientation orient);
-    LOCAL_REGIONS_EXPORT virtual void v_GetEdgeInterpVals(
-        const int edge, const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, NekDouble> &outarray,
+        StdRegions::Orientation orient) override;
     LOCAL_REGIONS_EXPORT virtual void v_GetTraceQFactors(
-        const int edge, Array<OneD, NekDouble> &outarray);
-    LOCAL_REGIONS_EXPORT virtual void v_ComputeTraceNormal(const int edge);
+        const int edge, Array<OneD, NekDouble> &outarray) override;
+    LOCAL_REGIONS_EXPORT virtual void v_ComputeTraceNormal(
+        const int edge) override;
 
     //---------------------------------------
     // Helper functions
     //---------------------------------------
-    LOCAL_REGIONS_EXPORT virtual int v_GetCoordim();
     LOCAL_REGIONS_EXPORT virtual void v_ExtractDataToCoeffs(
         const NekDouble *data, const std::vector<unsigned int> &nummodes,
         const int mode_offset, NekDouble *coeffs,
-        std::vector<LibUtilities::BasisType> &fromType);
+        std::vector<LibUtilities::BasisType> &fromType) override;
     LOCAL_REGIONS_EXPORT virtual StdRegions::Orientation v_GetTraceOrient(
-        int edge);
-    LOCAL_REGIONS_EXPORT virtual const LibUtilities::BasisSharedPtr &v_GetBasis(
-        int dir) const;
-    LOCAL_REGIONS_EXPORT virtual int v_GetNumPoints(const int dir) const;
+        int edge) override;
     LOCAL_REGIONS_EXPORT virtual void v_GetTracePhysMap(
-        const int edge, Array<OneD, int> &outarray);
+        const int edge, Array<OneD, int> &outarray) override;
 
     //---------------------------------------
     // Matrix creation functions
     //---------------------------------------
     LOCAL_REGIONS_EXPORT virtual DNekMatSharedPtr v_GenMatrix(
-        const StdRegions::StdMatrixKey &mkey);
+        const StdRegions::StdMatrixKey &mkey) override;
     LOCAL_REGIONS_EXPORT virtual DNekMatSharedPtr v_CreateStdMatrix(
-        const StdRegions::StdMatrixKey &mkey);
+        const StdRegions::StdMatrixKey &mkey) override;
     LOCAL_REGIONS_EXPORT virtual DNekScalMatSharedPtr v_GetLocMatrix(
-        const MatrixKey &mkey);
+        const MatrixKey &mkey) override;
     LOCAL_REGIONS_EXPORT virtual DNekScalBlkMatSharedPtr v_GetLocStaticCondMatrix(
-        const MatrixKey &mkey);
+        const MatrixKey &mkey) override;
 
-    LOCAL_REGIONS_EXPORT void v_DropLocStaticCondMatrix(const MatrixKey &mkey);
+    LOCAL_REGIONS_EXPORT void v_DropLocStaticCondMatrix(
+        const MatrixKey &mkey) override;
 
     LOCAL_REGIONS_EXPORT virtual void v_MassMatrixOp(
         const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray, const StdRegions::StdMatrixKey &mkey);
+        Array<OneD, NekDouble> &outarray,
+        const StdRegions::StdMatrixKey &mkey) override;
     LOCAL_REGIONS_EXPORT virtual void v_LaplacianMatrixOp(
         const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray, const StdRegions::StdMatrixKey &mkey);
+        Array<OneD, NekDouble> &outarray,
+        const StdRegions::StdMatrixKey &mkey) override;
     LOCAL_REGIONS_EXPORT virtual void v_LaplacianMatrixOp(
         const int k1, const int k2, const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray, const StdRegions::StdMatrixKey &mkey);
+        Array<OneD, NekDouble> &outarray,
+        const StdRegions::StdMatrixKey &mkey) override;
     LOCAL_REGIONS_EXPORT virtual void v_WeakDerivMatrixOp(
         const int i, const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray, const StdRegions::StdMatrixKey &mkey);
+        Array<OneD, NekDouble> &outarray,
+        const StdRegions::StdMatrixKey &mkey) override;
     LOCAL_REGIONS_EXPORT virtual void v_WeakDirectionalDerivMatrixOp(
         const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray, const StdRegions::StdMatrixKey &mkey);
+        Array<OneD, NekDouble> &outarray,
+        const StdRegions::StdMatrixKey &mkey) override;
     LOCAL_REGIONS_EXPORT virtual void v_MassLevelCurvatureMatrixOp(
         const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray, const StdRegions::StdMatrixKey &mkey);
+        Array<OneD, NekDouble> &outarray,
+        const StdRegions::StdMatrixKey &mkey) override;
     LOCAL_REGIONS_EXPORT virtual void v_HelmholtzMatrixOp(
         const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray, const StdRegions::StdMatrixKey &mkey);
-    LOCAL_REGIONS_EXPORT virtual void v_GeneralMatrixOp_MatOp(
-        const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray, const StdRegions::StdMatrixKey &mkey);
+        Array<OneD, NekDouble> &outarray,
+        const StdRegions::StdMatrixKey &mkey) override;
     LOCAL_REGIONS_EXPORT virtual void v_LaplacianMatrixOp_MatFree_Kernel(
         const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray, Array<OneD, NekDouble> &wsp);
+        Array<OneD, NekDouble> &outarray, Array<OneD, NekDouble> &wsp) override;
     LOCAL_REGIONS_EXPORT virtual void v_ReduceOrderCoeffs(
         int numMin, const Array<OneD, const NekDouble> &inarray,
-        Array<OneD, NekDouble> &outarray);
+        Array<OneD, NekDouble> &outarray) override;
 
-    LOCAL_REGIONS_EXPORT virtual void v_ComputeLaplacianMetric();
+    LOCAL_REGIONS_EXPORT virtual void v_ComputeLaplacianMetric() override;
 
     LOCAL_REGIONS_EXPORT virtual void v_SVVLaplacianFilter(
-        Array<OneD, NekDouble> &array, const StdRegions::StdMatrixKey &mkey);
+        Array<OneD, NekDouble> &array,
+        const StdRegions::StdMatrixKey &mkey) override;
 
     LOCAL_REGIONS_EXPORT virtual void v_NormalTraceDerivFactors(
         Array<OneD, Array<OneD, NekDouble>> &factors,
         Array<OneD, Array<OneD, NekDouble>> &d0factors,
-        Array<OneD, Array<OneD, NekDouble>> &d1factors);
+        Array<OneD, Array<OneD, NekDouble>> &d1factors) override;
 
 private:
     LibUtilities::NekManager<MatrixKey, DNekScalMat, MatrixKey::opLess>